requirements-dev.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
click==6.7
coverage==4.4.1
decorator==4.0.11
flake8==3.3.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mccabe==0.6.1
pexpect==4.2.1
pickleshare==0.7.4
prequ==1.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
pygments==2.2.0
pytest==3.1.2
pytest-cov==2.5.1
pytest-django==3.1.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
